Add sin,cos,tan
コンパイル時sin,cos,tanを追加

+++ saki/array/details/array_operator.h
@@ -7,16 +7,17 @@
 #pragma once
 #ifndef SAKI_ARRAY_OPERATOR_2018_12_24
 #define SAKI_ARRAY_OPERATOR_2018_12_24
+#include <cstddef>
 
 namespace saki
 {
 	//プロトタイプ宣言
-	template<typename T, unsigned int Size>
+	template<typename T, size_t Size>
 	class Array;
 	/**
 	* @brief ==演算子
 	*/
-	template<typename T, unsigned int Size>
+	template<typename T, size_t Size>
 	constexpr bool operator==(const Array<T, Size>& arr1, const Array<T, Size>& arr2)
 	{
 		for (unsigned int i = 0; i < Size; ++i)
@@ -28,7 +29,7 @@ namespace saki
 	/**
 	* @brief !=演算子
 	*/
-	template<typename T, unsigned int Size>
+	template<typename T, size_t Size>
 	constexpr bool operator!=(const Array<T, Size>& arr1, const Array<T, Size>& arr2)
 	{
 		return !(arr1 == arr2);
